refactor(interfaces): Add utility types
