--- modules/interfaces/src/key.ts
@@ -1,5 +1,0 @@
-/**
- * Better alternative to "keyof" in TypeScript.
- * Key of T and it's all strings.
- */
-export type Key<T> = Extract<keyof T, string>;
